// libs/features/complaints/src/lib/shared/models/establishment-summary.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';

export class EstablishmentSummary {
  establishmentName: BilingualText = new BilingualText();
  registrationNo: number;
  establishmentContactNumber: string;
  establishmentEmailId: string;
}